# [no ticket][risk=no] Improve e2e cohort tests