--- a/e2e/tests/cohorts/cohort-edit.spec.ts
+++ b/e2e/tests/cohorts/cohort-edit.spec.ts
@@ -1,27 +1,63 @@
+import { Page } from 'puppeteer';
 import WorkspaceDataPage from 'app/page/workspace-data-page';
 import { findOrCreateWorkspace, signInWithAccessToken } from 'utils/test-utils';
 import CohortBuildPage from 'app/page/cohort-build-page';
-import { makeRandomName, makeWorkspaceName } from 'utils/str-utils';
-import { waitWhileLoading } from 'utils/waits-utils';
+import { makeWorkspaceName } from 'utils/str-utils';
 import { MenuOption, ResourceCard } from 'app/text-labels';
 import DataResourceCard from 'app/component/data-resource-card';
 import Link from 'app/element/link';
+import { Visits } from 'app/page/cohort-participants-group';
 
-describe('Editing Cohort tests', () => {
+describe('Editing Cohort Test', () => {
+  // Tests require one and same workspace
+  const workspaceName = makeWorkspaceName();
+  let workspaceUrl: string;
+  let cohortName: string;
+
   beforeEach(async () => {
     await signInWithAccessToken(page);
   });
 
-  // Tests require one and same workspace
-  const workspace = makeWorkspaceName();
-  let workspaceUrl;
+  // Include all visit types in a single Include Group.
+  test('Create Cohort from all visits', async () => {
+    await loadWorkspaceUrl(page, workspaceName);
 
-  test('Discard changes', async () => {
-    await findAWorkspace();
+    // Add new Cohort.
+    const cohortBuildPage = await new WorkspaceDataPage(page).clickAddCohortsButton();
 
+    const group1 = cohortBuildPage.findIncludeParticipantsGroup('Group 1');
+    await group1.includeVisits([
+      Visits.AmbulanceVisit,
+      Visits.AmbulatoryClinicCenter,
+      Visits.AmbulatoryRehabilitationVisit,
+      Visits.EmergencyRoomVisit,
+      Visits.EmergencyRoomAndInpatientVisit,
+      Visits.HomeVisit,
+      Visits.InpatientVisit,
+      Visits.LaboratoryVisit,
+      Visits.NonhospitalInstitutionVisit,
+      Visits.OfficeVisit,
+      Visits.OutpatientVisit,
+      Visits.PharmacyVisit
+    ]);
+    const group1Count = await group1.getGroupCount();
+    const totalCount = await cohortBuildPage.getTotalCount();
+    expect(group1Count).toEqual(totalCount);
+    expect(Number.isNaN(group1Count)).toBe(false);
+
+    // Save new cohort.
+    cohortName = await cohortBuildPage.createCohort();
+
+    // Delete cohort in Cohort Build page.
+    // await new CohortActionsPage(page).deleteCohort();
+  });
+
+  test('Discard changes during editing', async () => {
+    await loadWorkspaceUrl(page, workspaceName);
+
+    // Open previously created Cohort.
     const dataPage = new WorkspaceDataPage(page);
-    const cohortCard = await dataPage.findOrCreateCohort();
-    const cohortName = await cohortCard.getResourceName();
+    const cohortCard = await dataPage.findCohortCard(cohortName);
     await cohortCard.clickResourceName();
 
     const cohortBuildPage = new CohortBuildPage(page);
@@ -30,6 +66,7 @@ describe('Editing Cohort tests', () => {
     // Wait for Total Count that also indicates page load is ready.
     await cohortBuildPage.getTotalCount();
 
+    // Start making some changes. All changes won't be saved.
     // Switch on Temporal in Group 1.
     const group1 = cohortBuildPage.findIncludeParticipantsGroupByIndex();
     await group1.clickTemporalSwitch(true);
@@ -39,34 +76,17 @@ describe('Editing Cohort tests', () => {
 
     await cohortBuildPage.discardChangesConfirmationDialog();
 
-    // Back to Data page.
+    // Discard changes brings user back to the Data page.
     await dataPage.waitForLoad();
-    const resourceCard = new DataResourceCard(page);
-    expect(await resourceCard.findCard(cohortName, ResourceCard.Cohort)).toBeTruthy();
   });
 
-  test('Delete cohort', async () => {
-    const cohortName = await setUpWorkspaceAndCohort().then((cohort: DataResourceCard) => cohort.clickResourceName());
+  test('Save As another new cohort', async () => {
+    await loadWorkspaceUrl(page, workspaceName);
 
-    const cohortBuildPage = new CohortBuildPage(page);
-    await cohortBuildPage.waitForLoad();
-
-    // Delete cohort while inside the Cohort Build page
-    const modalContent = await cohortBuildPage.deleteCohort();
-    expect(modalContent).toContain(`Are you sure you want to delete Cohort: ${cohortName}?`);
-
-    // Back to the Data page.
+    // Open previously created Cohort.
     const dataPage = new WorkspaceDataPage(page);
-    await dataPage.waitForLoad();
-
-    const resourceCard = new DataResourceCard(page);
-    expect(await resourceCard.findCard(cohortName, ResourceCard.Cohort)).toBeFalsy();
-  });
-
-  test('Save as cohort', async () => {
-    const originalCohortName = await setUpWorkspaceAndCohort().then((cohort: DataResourceCard) =>
-      cohort.clickResourceName()
-    );
+    const cohortCard = await dataPage.findCohortCard(cohortName);
+    await cohortCard.clickResourceName();
 
     const cohortBuildPage = new CohortBuildPage(page);
     await cohortBuildPage.waitForLoad();
@@ -75,11 +95,11 @@ describe('Editing Cohort tests', () => {
     const totalCount = await cohortBuildPage.getTotalCount();
 
     // Save as.
-    const cohortName = await cohortBuildPage.saveChanges(MenuOption.SaveAs);
-    console.log(`Saved as Cohort: "${cohortName}"`);
+    const newCohortName = await cohortBuildPage.saveChanges(MenuOption.SaveAs);
+    console.log(`Saved as Cohort: "${newCohortName}"`);
 
     // Click new cohort name link. Open Cohort Build page.
-    const cohortLink = Link.findByName(page, { name: cohortName });
+    const cohortLink = Link.findByName(page, { name: newCohortName });
     await cohortLink.clickAndWait();
 
     // Total Count should be unchanged.
@@ -87,74 +107,26 @@ describe('Editing Cohort tests', () => {
 
     // Delete cohort while inside the Cohort Build page
     const modalContent = await cohortBuildPage.deleteCohort();
-    expect(modalContent).toContain(`Are you sure you want to delete Cohort: ${cohortName}?`);
+    expect(modalContent).toContain(`Are you sure you want to delete Cohort: ${newCohortName}?`);
 
     // Back to the Data page.
-    const dataPage = new WorkspaceDataPage(page);
     await dataPage.waitForLoad();
 
     const resourceCard = new DataResourceCard(page);
     // Save as cohort is gone.
-    expect(await resourceCard.findCard(cohortName, ResourceCard.Cohort)).toBeFalsy();
+    expect(await resourceCard.findCard(newCohortName, ResourceCard.Cohort)).toBeFalsy();
     // Original cohort remains.
-    expect(await resourceCard.findCard(originalCohortName, ResourceCard.Cohort)).toBeTruthy();
+    expect(await resourceCard.findCard(cohortName, ResourceCard.Cohort)).toBeTruthy();
   });
 
   // Helper functions
-  async function findAWorkspace(): Promise<void> {
+  async function loadWorkspaceUrl(page: Page, workspaceName: string): Promise<void> {
     if (workspaceUrl) {
-      // Faster: Load previously saved URL instead clicks thru pages to open workspace data page.
-      await page.goto(workspaceUrl, { waitUntil: ['load', 'domcontentloaded', 'networkidle0'] });
+      // Faster: Load previously saved URL instead clicks thru links to open workspace data page.
+      await page.goto(workspaceUrl, { waitUntil: ['load', 'networkidle0'] });
       return;
     }
-    await findOrCreateWorkspace(page, { workspaceName: workspace });
-    workspaceUrl = page.url();
-  }
-
-  async function setUpWorkspaceAndCohort(): Promise<DataResourceCard> {
-    await findAWorkspace();
-    await deleteAllCohort('Duplicate');
-    const dataPage = new WorkspaceDataPage(page);
-    const cohortCard = await dataPage.findOrCreateCohort();
-    return duplicateCohort(cohortCard);
-  }
-
-  // Make a duplicate of an existing cohort to avoid hitting issue of concurrent data editing.
-  async function duplicateCohort(cohortCard: DataResourceCard): Promise<DataResourceCard> {
-    const cohortName = await cohortCard.getResourceName();
-    const snowmanMenu = await cohortCard.getSnowmanMenu();
-    await snowmanMenu.select(MenuOption.Duplicate, { waitForNav: false });
-    await waitWhileLoading(page);
-    const duplicateCohortName = `Duplicate of ${cohortName}`;
-    const dataPage = new WorkspaceDataPage(page);
-    // Rename duplicate Cohort name
-    const newCohortName = makeRandomName();
-    await dataPage.renameResource(duplicateCohortName, newCohortName, ResourceCard.Cohort);
-    // Verify rename successful.
-    expect(await DataResourceCard.findCard(page, newCohortName)).toBeTruthy();
-    expect(await DataResourceCard.findCard(page, cohortName)).toBeTruthy();
-    return dataPage.findCohortCard(newCohortName);
-  }
-
-  // Delete existing Cohorts with name contains 'Duplicate' string.
-  async function deleteAllCohort(pattern: string): Promise<void> {
-    const dataResourceCard = new DataResourceCard(page);
-    const cards = await dataResourceCard.getResourceCard(ResourceCard.Cohort);
-
-    const cardNames = [];
-    for (const card of cards) {
-      const name = await card.getResourceName();
-      if (name.includes(pattern)) {
-        cardNames.push(await card.getResourceName());
-      }
-    }
-
-    for (const name of cardNames) {
-      const modalTextContent = await new WorkspaceDataPage(page).deleteResource(name, ResourceCard.Cohort);
-      // Verify Delete Confirmation modal text
-      expect(modalTextContent).toContain(`Are you sure you want to delete Cohort: ${name}?`);
-      expect(await DataResourceCard.findCard(page, name, 2000)).toBeFalsy();
-      console.log(`deleted cohort ${name}`);
-    }
+    await findOrCreateWorkspace(page, { workspaceName });
+    workspaceUrl = page.url(); // Save URL for load workspace directly without search.
   }
 });