--- a/e2e/utils/test-utils.ts
+++ b/e2e/utils/test-utils.ts
@@ -131,7 +131,7 @@ export async function createWorkspace(
   const workspacesPage = new WorkspacesPage(page);
   await workspacesPage.load();
   await workspacesPage.createWorkspace(workspaceName, cdrVersionName);
-  console.log(`Created workspace "${workspaceName}" with CDR version "${cdrVersionName}"`);
+  logger.info(`Created workspace "${workspaceName}" with CDR version "${cdrVersionName}"`);
   return workspaceName;
 }