Added new teset `Cohort search by code and description` in `cohort-ui.spec` and reorganized `cohort-edit.spec` and `cohort-create.spec` tests to make them run faster.

Sample run times:

Before:
`PASS tests/cohorts/cohort-edit.spec.ts (174.943 s)`
After:
`PASS tests/cohorts/cohort-edit.spec.ts (110.571 s)`

Before:
`PASS tests/cohorts/cohort-create.spec.ts (380.615 s)`
After:
`PASS tests/cohorts/cohort-create.spec.ts (316.047 s)`